Moved type attribute from SourceDatasetStats to SourceDataset. SourceDatasetType is indicated in the source dataset lists using a style.
