+++ publisher-domain/src/main/java/nl/idgis/publisher/domain/web/SourceDatasetStats.java
@@ -1,7 +1,5 @@
 package nl.idgis.publisher.domain.web;
 
-import nl.idgis.publisher.domain.SourceDatasetType;
-
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonGetter;
 import com.fasterxml.jackson.annotation.JsonProperty;
@@ -12,21 +10,15 @@ public class SourceDatasetStats extends Entity {
 	
 	private final SourceDataset 	sourceDataset;
 	private final long				datasetCount;
-	private final SourceDatasetType	type;
 	
 	@JsonCreator
 	public SourceDatasetStats (
 			final @JsonProperty("sourceDataset") SourceDataset sourceDataset, 
-			final @JsonProperty("datasetCount") long datasetCount,
-			final @JsonProperty("type") SourceDatasetType type) {
+			final @JsonProperty("datasetCount") long datasetCount) {
 		
-		if (type == null) {
-			throw new NullPointerException ("type cannot be null");
-		}
 		
 		this.sourceDataset = sourceDataset;
 		this.datasetCount = datasetCount;
-		this.type = type;
 	}
 
 	@JsonGetter
@@ -37,10 +29,5 @@ public class SourceDatasetStats extends Entity {
 	@JsonGetter
 	public long datasetCount () {
 		return this.datasetCount;
-	}
-	
-	@JsonGetter
-	public SourceDatasetType type () {
-		return this.type;
 	}
 }
