+++ publisher-web/app/views/datasources/list.scala.html
@@ -5,6 +5,7 @@
 	currentCategory: nl.idgis.publisher.domain.web.Category,
 	search: String)
 
+@import nl.idgis.publisher.domain._
 @import nl.idgis.publisher.domain.web._
 
 @dataSourceLink(ds: DataSource, cls: String = null) = {
@@ -157,7 +158,7 @@
 
 					<div class="list-group">
 						@for(dataset <- sourceDatasets.values) {
-							<div class="list-group-item">
+							<div class="list-group-item@if(dataset.sourceDataset.`type` == SourceDatasetType.UNAVAILABLE){ list-group-item-warning}">
 								<div class="row">
 									<div class="col-md-5">
 										<h4 class="list-group-item-heading">@dataset.sourceDataset.name</h4>
@@ -180,7 +181,9 @@
 										</p>
 									</div>
 									<div class="col-md-2 text-right">
-										<a class="btn btn-success" href="@routes.Datasets.createFormForSourceDataset(dataset.sourceDataset.id)" title="@Domain.message("web.application.page.sourcedatasets.list.button.newdataset")"><span class="glyphicon glyphicon-plus"></span></a>
+										@if(dataset.sourceDataset.`type` != SourceDatasetType.UNAVAILABLE) {
+											<a class="btn btn-success" href="@routes.Datasets.createFormForSourceDataset(dataset.sourceDataset.id)" title="@Domain.message("web.application.page.sourcedatasets.list.button.newdataset")"><span class="glyphicon glyphicon-plus"></span></a>
+										}
 									</div>
 								</div>
 							</div>
